fix: 퇴장룸

See merge request s09-webmobile1-sub2/S09P12C208!99

+++ Backend/src/main/java/com/narang_norang/NarangNorang/redis/picture/domain/dto/PictureResponse.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class PictureResponse {
 
-    private Long pictureSeq;
+    private Integer pictureSeq;
 
     private String roomCode;
